Move the caret-handling stubs into `CaretControl`.
